a big mistake: the return type in the "factorial" function prototype is different from the function definition.

+++ Chap01/Example1_3/factorial.cpp
@@ -1,4 +1,4 @@
-int factorial(long n)
+long factorial(long n)
 {
     long ans(1);
     for (long i=2; i<=n; ++i ){
