+++ Chap01/Example1_3/factorial.h
@@ -1,6 +1,6 @@
 #ifndef FACTORIAL_H
 #define FACTORIAL_H
 
-int factorial(long n);
+long factorial(long n);
 
 #endif // FACTORIAL_H
